requirements/base.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dataclasses-json==0.6.2
distro==1.8.0
emoji==2.8.0
exceptiongroup==1.1.3
filetype==1.2.0
frozenlist==1.4.0
gidgethub==5.3.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.335
langdetect==1.0.9
langsmith==0.0.64
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.2
openai==1.2.4
packaging==23.2
pluggy==1.3.0
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tabulate==0.9.0
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.10.30
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.2
